feedback: Extract ScalarAdminClientFactory interface to domain layer (DIP)
Move the factory abstraction to domain/client/ so that PauseApplicationService
no longer depends on the infrastructure layer, consistent with how
KubernetesClient/KubernetesClientImpl is structured.
