+++ lib/src/main/java/com/scalar/admin/kubernetes/application/PauseApplicationService.java
@@ -9,7 +9,7 @@ import com.scalar.admin.kubernetes.domain.model.pause.PauseDuration;
 import com.scalar.admin.kubernetes.domain.model.pause.PauseTarget;
 import com.scalar.admin.kubernetes.domain.client.KubernetesClient;
 import com.scalar.admin.kubernetes.domain.service.PauseService;
-import com.scalar.admin.kubernetes.infrastructure.client.ScalarAdminClientFactory;
+import com.scalar.admin.kubernetes.domain.client.ScalarAdminClientFactory;
 import javax.annotation.concurrent.NotThreadSafe;
 
 /**
