+++ lib/src/main/java/com/scalar/admin/kubernetes/infrastructure/client/ScalarAdminClientFactoryImpl.java
@@ -1,18 +1,20 @@
 package com.scalar.admin.kubernetes.infrastructure.client;
 
 import com.scalar.admin.kubernetes.domain.client.ScalarAdminClient;
+import com.scalar.admin.kubernetes.domain.client.ScalarAdminClientFactory;
 import com.scalar.admin.kubernetes.domain.model.pause.PauseTarget;
 import com.scalar.admin.kubernetes.domain.model.pause.TlsConfig;
 import javax.annotation.concurrent.ThreadSafe;
 
 /**
- * Factory for creating ScalarAdminClient instances.
+ * Implementation of {@link ScalarAdminClientFactory} using concrete ScalarAdminClient
+ * implementations.
  *
  * <p>This factory creates appropriate ScalarAdminClient implementations based on whether TLS
  * configuration is provided. Each client is bound to a specific PauseTarget.
  */
 @ThreadSafe
-public class ScalarAdminClientFactory {
+public class ScalarAdminClientFactoryImpl implements ScalarAdminClientFactory {
 
   /**
    * Creates a standard (non-TLS) ScalarAdminClient for the given target.
@@ -20,6 +22,7 @@ public class ScalarAdminClientFactory {
    * @param target the pause target containing pods to communicate with
    * @return a new ScalarAdminClient instance without TLS
    */
+  @Override
   public ScalarAdminClient createClient(PauseTarget target) {
     return new ScalarAdminClientImpl(target);
   }
@@ -31,6 +34,7 @@ public class ScalarAdminClientFactory {
    * @param tlsConfig the TLS configuration for secure communication
    * @return a new ScalarAdminClient instance with TLS enabled
    */
+  @Override
   public ScalarAdminClient createClient(PauseTarget target, TlsConfig tlsConfig) {
     return new TlsScalarAdminClientImpl(target, tlsConfig);
   }
